==> P1.r <==
#D1-preprocessing data: Outliers&Correlations
setwd("X:/Data/Desktop/Arabidopsis Arrays/Original Data")
arr <- read.table("exp_ann30K.txt")
dim(arr)
arr <- arr[1:29999,17:196]
#Boxplot-Outliers(operations)
boxplot(arr[,1:8], range = 0.5)
boxplot(arr[,173:180], range = 0.5)
png(filename = "Boxplot-RIL.png", width = 6720, height = 3780, bg = "white")
boxplot(arr[,9:172], range = 1)
dev.off()
#Cor&heatmap-Correlation(identical)
image(cor(arr[,1:8]))
image(cor(arr[,173:180]))
image(cor(arr[,9:172]))
heatmap(cor(arr[,1:8]),scale="none")
png(filename = "HM_Cor-RIL.png", width = 1920, height = 1920, bg = "white")
heatmap(cor(arr[,9:172]),scale="none")
dev.off()
heatmap(cor(arr[,173:180]),scale="none")
#Normalization---with package limma
arrQ <- normalizeQuantiles(arr)
boxplot(arrQ[,1:8], range = 0.5)
boxplot(arrQ[,173:180], range = 0.5)
png(filename = "Boxplot-RIL-Q.png", width = 6720, height = 3780, bg = "white")
boxplot(arrQ[,9:172], range = 1)
dev.off()
image(cor(arrQ[,1:8]))
image(cor(arrQ[,173:180]))
image(cor(arrQ[,9:172]))
heatmap(cor(arrQ[,1:8]),scale="none")
png(filename = "HM_Cor-RIL-Q.png", width = 1920, height = 1920, bg = "white")
heatmap(cor(arrQ[,9:172]),scale="none")
dev.off()
heatmap(cor(arrQ[,173:180]),scale="none")
